+++ Sources/Player/Player/Player+Metrics.swift
@@ -25,6 +25,17 @@ public extension Player {
             .eraseToAnyPublisher()
     }
 
+    /// The current player metrics, if available.
+    ///
+    /// Each call to this function might return different results reflecting the most recent metrics available. The
+    /// included ``Metrics/increment`` collates data from the entire playback session and is therefore always equal
+    /// to ``Metrics/total``.
+    ///
+    /// > Important: Metrics are reset when toggling external playback.
+    func metrics() -> Metrics? {
+        properties.metrics()
+    }
+
     /// Returns a periodically updated metrics history associated with the current item.
     ///
     /// - Parameters:
