+++ Sources/Player/Player/Player+Playback.swift
@@ -5,6 +5,21 @@
 //
 
 public extension Player {
+    /// A Boolean value whether the player should play content when possible.
+    var shouldPlay: Bool {
+        get {
+            queuePlayer.rate != 0
+        }
+        set {
+            if newValue {
+                queuePlayer.rate = queuePlayer.defaultRate
+            }
+            else {
+                queuePlayer.rate = 0
+            }
+        }
+    }
+
     /// Resumes playback.
     func play() {
         shouldPlay = true
